+++ api/revanced-patches.api
@@ -301,24 +301,6 @@ public final class app/revanced/patches/lightroom/misc/premium/UnlockPremiumPatc
 	public synthetic fun execute (Lapp/revanced/patcher/data/Context;)V
 }
 
-public final class app/revanced/patches/magazines/misc/customtabs/UseCustomTabs : app/revanced/patcher/patch/BytecodePatch {
-	public static final field INSTANCE Lapp/revanced/patches/magazines/misc/customtabs/UseCustomTabs;
-	public fun execute (Lapp/revanced/patcher/data/BytecodeContext;)V
-	public synthetic fun execute (Lapp/revanced/patcher/data/Context;)V
-}
-
-public final class app/revanced/patches/magazines/misc/gms/GmsCoreSupportPatch : app/revanced/patches/shared/misc/gms/BaseGmsCoreSupportPatch {
-	public static final field INSTANCE Lapp/revanced/patches/magazines/misc/gms/GmsCoreSupportPatch;
-}
-
-public final class app/revanced/patches/magazines/misc/gms/GmsCoreSupportResourcePatch : app/revanced/patches/shared/misc/gms/BaseGmsCoreSupportResourcePatch {
-	public static final field INSTANCE Lapp/revanced/patches/magazines/misc/gms/GmsCoreSupportResourcePatch;
-}
-
-public final class app/revanced/patches/magazines/misc/integrations/IntegrationsPatch : app/revanced/patches/shared/misc/integrations/BaseIntegrationsPatch {
-	public static final field INSTANCE Lapp/revanced/patches/magazines/misc/integrations/IntegrationsPatch;
-}
-
 public final class app/revanced/patches/memegenerator/detection/license/LicenseValidationPatch : app/revanced/patcher/patch/BytecodePatch {
 	public static final field INSTANCE Lapp/revanced/patches/memegenerator/detection/license/LicenseValidationPatch;
 	public fun execute (Lapp/revanced/patcher/data/BytecodeContext;)V
