--- src/main/kotlin/app/revanced/patches/magazines/misc/gms/Constants.kt
@@ -1,6 +1,0 @@
-package app.revanced.patches.magazines.misc.gms
-
-internal object Constants {
-    const val MAGAZINES_PACKAGE_NAME = "com.google.android.apps.magazines"
-    const val REVANCED_MAGAZINES_PACKAGE_NAME = "app.revanced.android.magazines"
-}
