--- src/main/kotlin/app/revanced/patches/magazines/misc/gms/fingerprints/GooglePlayUtilityFingerprint.kt
@@ -1,12 +1,0 @@
-package app.revanced.patches.magazines.misc.gms.fingerprints
-
-import app.revanced.patcher.extensions.or
-import app.revanced.patcher.fingerprint.MethodFingerprint
-import com.android.tools.smali.dexlib2.AccessFlags
-
-internal object GooglePlayUtilityFingerprint : MethodFingerprint(
-    returnType = "I",
-    accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
-    parameters = listOf("L", "I"),
-    strings = listOf("This should never happen.", "MetadataValueReader", "com.google.android.gms")
-)
