--- src/main/kotlin/app/revanced/patches/magazines/misc/gms/fingerprints/MagazinesActivityOnCreateFingerprint.kt
@@ -1,9 +1,0 @@
-package app.revanced.patches.magazines.misc.gms.fingerprints
-
-import app.revanced.patcher.fingerprint.MethodFingerprint
-
-internal object MagazinesActivityOnCreateFingerprint : MethodFingerprint(
-    customFingerprint = { methodDef, classDef ->
-       classDef.type.endsWith("/StartActivity;") && methodDef.name == "onCreate"
-    }
-)
