--- src/main/kotlin/app/revanced/patches/magazines/misc/gms/fingerprints/ServiceCheckFingerprint.kt
@@ -1,12 +1,0 @@
-package app.revanced.patches.magazines.misc.gms.fingerprints
-
-import app.revanced.patcher.extensions.or
-import app.revanced.patcher.fingerprint.MethodFingerprint
-import com.android.tools.smali.dexlib2.AccessFlags
-
-internal object ServiceCheckFingerprint : MethodFingerprint(
-    returnType = "V",
-    accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
-    parameters = listOf("L", "I"),
-    strings = listOf("Google Play Services not available", "GooglePlayServices not available due to error ")
-)
